requirements.txt
tensorflow
numpy
grpcio
tensorflow-hub
pytest
pillow